requirements/base.txt
django==5.0.3
psycopg2-binary==2.9.9
python-environ==0.4.54
django-cors-headers==4.3.1
django-jazzmin==2.6.1
drf-yasg==1.21.7
djangorestframework==3.15.1
django-filter==24.2
django-modeltranslation==0.18.11
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
Pillow==10.2.0
eskiz-sms==0.2.3
# firebase
fcm-django==2.0.1

# testing
factory-boy==3.2.1
model_bakery==1.11.0
pytest==7.3.1
#pytest-mock==3.11.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
#pytest-env==0.8.2
#pytest-xdist==3.3.1

# phone number
django-phonenumber-field==7.1.0
phonenumbers==8.13.19


# ocpp service
ocpp==0.26.0
aiohttp==3.9.3
fastapi==0.110.1
uvicorn==0.29.0
pandantic==0.3.0
wsproto==1.2.0
websocket==0.2.1


# websocket
django-channels==0.7.0
daphne==4.1.0
channels_redis==4.1.0

# caches
django-redis==5.2.0
redis==4.5.3

# sentry integration and logging
sentry-sdk==1.45.0
python-telegram-bot==13.15

# captcha
django-recaptcha==3.0.0


# celery
celery==5.4.0

django-rosetta==0.10.0